Don't sent wrong headers when making bulk imports
When making bulk imports with `import_users`, the default `Content-Type` headers (`application/json`)
should not be used, since there is a file in the payload.

The RestClient client fixes that and replaces this header with its own (thank you RestClient), but it
emits a warning, and it's still a mistake.

Since generating headers for a multipart POST request is a bit annoying, I've chosen to simply _remove_
the `Content-Type` headers, and let RestClient do the work.