Change strategy when normalizing path
When encoding the uri, we currently have two different strategies, either the full URL or the path.  When passing in just the path, the Addressable aggresively parses the path assuming its a full url.  This causes some charachters to not be encoded and instead be parsing points of the url.

This commit changes the strategy to check for when the base_uri is set, and if is is we dont need to parse the url, just encode the path